fix: `Conditions` element position
The order of elements in XML is significant, like the order of objects
in an array.
When a document is consumed, the order of elements may be irrelevant to
the consumer.
When the schema of a document is validated, i.e. with XSD, the order of
elements has to match the schema document.

The published standard XSD for SAML assertions has the `Conditions`
element before the `AudienceRestriction` element.
https://docs.oasis-open.org/security/saml/v2.0/saml-schema-assertion-2.0.xsd

This change moves the element in the default template to match.